refactor: moved gsap provider
