+++ src/context/billing/BillingContext.tsx
@@ -1,18 +1,15 @@
 'use client'
 
-import React, {
-  createContext,
-  useContext,
-  useState,
-  PropsWithChildren,
-} from 'react'
+import { createContext, useContext } from 'react'
 
 type BillingContextType = {
   isAnnual: boolean
   toggleBilling: () => void
 }
 
-const BillingContext = createContext<BillingContextType | undefined>(undefined)
+export const BillingContext = createContext<BillingContextType | undefined>(
+  undefined
+)
 
 export const useBilling = () => {
   const context = useContext(BillingContext)
@@ -20,16 +17,4 @@ export const useBilling = () => {
     throw new Error('useBilling must be used within a BillingProvider')
   }
   return context
-}
-
-export const BillingProvider = ({ children }: PropsWithChildren) => {
-  const [isAnnual, setIsAnnual] = useState(false)
-
-  const toggleBilling = () => setIsAnnual(prev => !prev)
-
-  return (
-    <BillingContext.Provider value={{ isAnnual, toggleBilling }}>
-      {children}
-    </BillingContext.Provider>
-  )
 }
