+++ src/context/billing/BillingProvider.tsx
@@ -1,0 +1,14 @@
+import { PropsWithChildren, useState } from 'react'
+import { BillingContext } from './BillingContext'
+
+export const BillingProvider = ({ children }: PropsWithChildren) => {
+  const [isAnnual, setIsAnnual] = useState(false)
+
+  const toggleBilling = () => setIsAnnual(prev => !prev)
+
+  return (
+    <BillingContext.Provider value={{ isAnnual, toggleBilling }}>
+      {children}
+    </BillingContext.Provider>
+  )
+}
